feat: 예약시간을 참조하는 예약 검증 추가

+++ src/main/java/roomescape/service/ReservationService.java
@@ -63,6 +63,9 @@ public class ReservationService {
 
     public void deleteReservationTime(final Long reservationTimeId) {
         checkReservationTimeExist(reservationTimeId);
+        if (reservationRepository.existByTimeId(reservationTimeId)) {
+            throw new IllegalArgumentException("예약에 포함된 시간 정보는 삭제할 수 없습니다.");
+        }
         reservationTimeRepository.deleteById(reservationTimeId);
     }
